Finished config implementation and handler refactor to properly trigger functions
Hypothetically, all requisite changes and have been implemented to enable the Discord bot with lambdas. The configs need to be updated with the Discord and ESPN specific values for the target league and server, but otherwise we should be good to test.

This commit implements config changes so that the bot can be run during regular and delayed weeks. Many Lambdas should be scaffolded with specific CRONs to work on their desired timelines. Manual modification and redeployment will be necessary for both delayed weeks AND daylight savings changes.